Tighten RON schema tests and extract phase emission
